devel build, build docs for loginserver

+++ loginserver/install_taserver.sh
@@ -5,7 +5,11 @@ set -ex
 
 TASERVER_RELEASE_TAG="wine-v0.0.22"
 
-# get taserver
-wget -q -O taserver.zip "https://github.com/chickenbellyfin/taserver/archive/refs/tags/$TASERVER_RELEASE_TAG.zip"
-unzip -q taserver.zip
-mv "taserver-$TASERVER_RELEASE_TAG" "taserver"
+# Get taserver
+if [ ! -f taserver.zip ]; then
+  wget -q -O taserver.zip "https://github.com/chickenbellyfin/taserver/archive/refs/tags/$TASERVER_RELEASE_TAG.zip"
+  unzip -q taserver.zip
+  mv "taserver-$TASERVER_RELEASE_TAG" "taserver"
+else
+  unzip -q taserver.zip
+fi
